chore: Remove first attempt repo for Challenge 2
