--- Challenge2/.github/workflows/ci.yml
@@ -1,25 +1,0 @@
-name: CI
-
-on:
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v2
-        with:
-          node-version: "10"
-
-      - name: Install dependencies
-        run: npm install
-
-      - name: Run tests
-        run: npm test
