--- Challenge2b/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-forta.config.json
