--- Challenge2b/.prettierrc
@@ -1,5 +1,0 @@
-{
-  "printWidth": 120,
-  "trailingComma": "es5",
-  "singleQuote": false
-}
