--- Challenge2b/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  testPathIgnorePatterns: ["dist"],
-};
