--- Challenge2b/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "forta-agent-starter",
-  "displayName": "Forta Agent Starter",
-  "version": "0.0.1",
-  "description": "Forta Agent Typescript starter project",
-  "longDescription": "Welcome to the Forta Agent JavaScript starter project!\n\nForta Network presents this starter project as an entry point for bot developers who want to integrate with the Forta Network ecosystem. As part of our commitment to empowering developers, this project provides a foundation for building powerful and secure bots that can interact with the Forta platform and take advantage of its cutting-edge capabilities.\n\nThe Forta Agent SDK and CLI tool offer a seamless experience for developers who wish to leverage the decentralized and trustless environment provided by the Forta Network. Whether you're interested in creating DeFi bots, monitoring smart contract transactions, or generating real-time alerts, this starter project sets you on the right path to achieve your goals.\n\nJoin the Forta Network community and learn more about our mission to revolutionize decentralized finance and blockchain analytics: https://forta.org/",
-  "repository": "https://github.com/forta-network/forta-bot-sdk/tree/master/starter-project",
-  "licenseUrl": "https://github.com/forta-network/forta-bot-sdk/blob/master/starter-project/LICENSE.md",
-  "promoUrl": "https://forta.org",
-  "format": "prettier --write \"src/**/*.ts\"",
-  "chainIds": [
-    1
-  ],
-  "scripts": {
-    "build": "tsc",
-    "start": "npm run start:dev",
-    "start:dev": "nodemon --watch src --watch forta.config.json -e js,ts,json  --exec \"npm run build && forta-agent run\"",
-    "start:prod": "forta-agent run --prod",
-    "tx": "npm run build && forta-agent run --tx",
-    "block": "npm run build && forta-agent run --block",
-    "range": "npm run build && forta-agent run --range",
-    "alert": "npm run build && forta-agent run --alert",
-    "sequence": "npm run build && forta-agent run --sequence",
-    "file": "npm run build && forta-agent run --file",
-    "publish": "forta-agent publish",
-    "info": "forta-agent info",
-    "logs": "forta-agent logs",
-    "push": "forta-agent push",
-    "disable": "forta-agent disable",
-    "enable": "forta-agent enable",
-    "keyfile": "forta-agent keyfile",
-    "stake": "forta-agent stake",
-    "test": "jest"
-  },
-  "dependencies": {
-    "forta-agent": "^0.1.48",
-    "lru-cache": "^11.0.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.10",
-    "@types/nodemon": "^1.19.0",
-    "forta-agent-tools": "^3.2.12",
-    "jest": "^29.7.0",
-    "nodemon": "^2.0.8",
-    "prettier": "^3.3.3",
-    "ts-jest": "^29.1.1",
-    "typescript": "^4.3.4"
-  }
-}
