--- Challenge2b/src/constants.ts
@@ -1,11 +1,0 @@
-export const UNI_SWAP_EVENT_ABI =
-  "event Swap (address indexed sender, address indexed recipient, int256 amount0, int256 amount1, uint160 sqrtPriceX96, uint128 liquidity, int24 tick)";
-
-export const UNI_POOL_FUNCTIONS_ABI = [
-  "function token0() external view returns (address)",
-  "function token1() external view returns (address)",
-  "function fee() external view returns (uint24)",
-];
-export const UNI_FACTORY_ADDRESS = "0x1F98431c8aD98523631AE4a59f267346ea31F984";
-export const UNI_INIT_CODE_HASH =
-  "0xe34f199b19b2b4f47f68442619d555527d244f78a3297ea89325f843f87b8b54";
